reconfig attempt 1
