+++ backend/routes/steamRoute.js
@@ -1,0 +1,35 @@
+// server requirements to run
+const express = require('express');
+const bodyParser = require('body-parser');
+const cors = require('cors');
+const path = require('path');
+const env = require('dotenv').config();
+const request = require('request');
+
+const{app, STEAM_WEB_API_KEY} = require("../db");
+
+//steam api post
+app.post('/api/getSteamGames', async (req, res) => 
+{
+  // incoming: userId, steamId
+  // outgoing: appId, playtime
+  const url = 'https://api.steampowered.com/IPlayerService/GetOwnedGames/v0001/?'
+              + 'key=' + steamWebApiKey + '&steamid=' + req.body.steamId;
+
+  request.get(url, function(steamHttpBody) {
+    res.setHeader('Content-Type', 'application/json');
+    res.send(steamHttpBody);
+  });
+});
+
+// Server static assets if in production
+if (process.env.NODE_ENV === 'production') 
+{
+  // Set static folder
+  app.use(express.static('frontend/build'));
+
+  app.get('*', (req, res) => 
+ {
+    res.sendFile(path.resolve(__dirname, 'frontend', 'build', 'index.html'));
+  });
+}
